Manages lifecycle of observable
- Adds methods to manage the lifecycle
