+++ domain/src/main/java/com/globant/equattrocchio/domain/GetLatestImagesUseCase.kt
@@ -1,10 +1,10 @@
 package com.globant.equattrocchio.domain
 
 import com.globant.equattrocchio.domain.models.ResultDomainInput
-import com.globant.equattrocchio.domain.service.ImagesServicese
+import com.globant.equattrocchio.domain.service.ImagesServices
 import io.reactivex.Observable
 
-class GetLatestImagesUseCase(private val imagesService: ImagesServicese) {
+class GetLatestImagesUseCase(private val imagesService: ImagesServices) {
     operator fun invoke(): Observable<ResultDomainInput> {
         return imagesService.getLatestImagesRx()
     }
